+++ rest_framework/serializers/validators.py
@@ -2,6 +2,10 @@
 Validators.
 
 """
+import re
+
+import six
+
 from rest_framework.serializers.exceptions import ValidationError
 
 
@@ -78,7 +82,7 @@ class MinLengthValidator(BaseValidator):
 
         """
         if len(value) < self.min_length:
-            raise ValidationError(self.message.format(dict(min_length=self.min_length)))
+            raise ValidationError(self.message.format(min_length=self.min_length))
 
 
 class MaxLengthValidator(BaseValidator):
@@ -108,7 +112,7 @@ class MaxLengthValidator(BaseValidator):
 
         """
         if len(value) > self.max_length:
-            raise ValidationError(self.message.format(dict(max_length=self.max_length)))
+            raise ValidationError(self.message.format(max_length=self.max_length))
 
 
 class MinValueValidator(BaseValidator):
@@ -138,7 +142,7 @@ class MinValueValidator(BaseValidator):
 
         """
         if value < self.min_value:
-            raise ValidationError(self.message.format(dict(min_value=self.min_value)))
+            raise ValidationError(self.message.format(min_value=self.min_value))
 
 
 class MaxValueValidator(BaseValidator):
@@ -168,4 +172,50 @@ class MaxValueValidator(BaseValidator):
 
         """
         if value > self.max_value:
-            raise ValidationError(self.message.format(dict(max_value=self.max_value)))
+            raise ValidationError(self.message.format(max_value=self.max_value))
+
+
+class RegexValidator(BaseValidator):
+    """
+    Validator for check regex raw.
+
+    """
+    regex = ''
+    message = 'Enter a valid value.'
+    inverse_match = False
+    flags = 0
+
+    def __init__(self, regex, inverse_match=None, flags=None, *args, **kwargs):
+        """
+        Validator for check regex raw.
+
+        :param str regex: Regex for check.
+        :param bool inverse_match: Reverse check result.
+        :param int flags: Flags for compile regular.
+
+        """
+        super().__init__(*args, **kwargs)
+        if regex is not None:
+            self.regex = regex
+        if inverse_match is not None:
+            self.inverse_match = inverse_match
+        if flags is not None:
+            self.flags = flags
+
+        if self.flags and not isinstance(self.regex, six.string_types):
+            raise TypeError("If the flags are set, regex must be a regular expression string.")
+
+        self.regex = re.compile(self.regex, self.flags)
+
+    def __call__(self, value):
+        """
+        Validate that the input contains a match for the regular expression
+        if inverse_match is False, otherwise raise ValidationError.
+
+        :param object value: Value for validation.
+
+        :raise: ValidationError: If not valid data.
+
+        """
+        if not (self.inverse_match is not bool(self.regex.search(value))):
+            raise ValidationError(self.message)